# Conflicts:
#	Projects/UnitTests/CMakeLists.txt
#	Sources/CMake/Modules/CMake-common.cmake
#	Sources/CMake/Modules/DavaTemplate.cmake
#	Sources/CMake/Modules/FindQT5.cmake
#	Sources/CMake/Modules/PlatformSettings.cmake
#	Sources/Internal/Autotesting/AutotestingSystem.cpp
#	Sources/Internal/Autotesting/AutotestingSystemLua.cpp
#	Sources/Internal/CMakeLists.txt
#	Sources/Internal/Collision/CollisionObject2.h
#	Sources/Internal/Core/ApplicationCore.h
#	Sources/Internal/Debug/DVAssertMessageApple.mm
#	Sources/Internal/Debug/Replay.cpp
#	Sources/Internal/FileSystem/VariantType.cpp
#	Sources/Internal/FileSystem/VariantType.h
#	Sources/Internal/Input/InputSystem.cpp
#	Sources/Internal/Input/KeyboardDevice.cpp
#	Sources/Internal/Math/Rect.h
#	Sources/Internal/Math/Vector.h
#	Sources/Internal/Particles/ParticleEmitter.cpp
#	Sources/Internal/Particles/ParticleLayer.cpp
#	Sources/Internal/Particles/ParticleLayer.h
#	Sources/Internal/Particles/ParticlePropertyLine.h
#	Sources/Internal/Particles/ParticleRenderObject.cpp
#	Sources/Internal/Platform/TemplateAndroid/CorePlatformAndroid.h
#	Sources/Internal/Platform/TemplateAndroid/ExternC/AndroidLayer.cpp
#	Sources/Internal/Platform/TemplateAndroid/ExternC/JNITextField.cpp
#	Sources/Internal/Platform/TemplateMacOS/MainWindowController.mm
#	Sources/Internal/Platform/TemplateMacOS/NPAPIPluginMacOS.mm
#	Sources/Internal/Platform/TemplateMacOS/OpenGLView.mm
#	Sources/Internal/Platform/TemplateWin32/CorePlatformWin32.cpp
#	Sources/Internal/Render/2D/FTFont.cpp
#	Sources/Internal/Render/2D/TextBlock.cpp
#	Sources/Internal/Render/3D/PolygonGroup.h
#	Sources/Internal/Render/Highlevel/Landscape.cpp
#	Sources/Internal/Render/Highlevel/RenderBatch.cpp
#	Sources/Internal/Render/Highlevel/RenderHierarchy.cpp
#	Sources/Internal/Render/Highlevel/RenderPass.cpp
#	Sources/Internal/Render/Highlevel/SpatialTree.cpp
#	Sources/Internal/Render/Highlevel/SpatialTree.h
#	Sources/Internal/Render/Highlevel/SpriteObject.h
#	Sources/Internal/Render/Image/LibDdsHelper.cpp
#	Sources/Internal/Render/Material/NMaterialNames.h
#	Sources/Internal/Render/PixelFormatDescriptor.h
#	Sources/Internal/Render/RHI/GLES2/rhi_DepthStencilStateGLES2.cpp
#	Sources/Internal/Render/RHI/GLES2/rhi_TextureGLES2.cpp
#	Sources/Internal/UI/UIControl.cpp
#	Sources/Internal/UI/UIJoypad.cpp
#	Sources/Internal/UI/UIScrollBar.cpp
#	Sources/Internal/UI/UIScrollView.h
#	Sources/Internal/UI/UIScrollViewContainer.cpp
#	Sources/Internal/UI/UISlider.cpp
#	Sources/Internal/UI/UITextFieldAndroid.h
#	Sources/Internal/UI/UITextFieldiPhone.h
#	Sources/Tools/QtTools/DavaGLWidget/davaglwidget.cpp
#	Sources/Tools/TexturePacker/TexturePacker.cpp
#	Tools/QuickEd/Classes/Document.cpp
#	Tools/ResourceEditor/CMakeLists.txt
#	Tools/ResourceEditor/Classes/Qt/DockParticleEditor/ParticleEditorWidget.cpp
#	Tools/ResourceEditor/Classes/Qt/Scene/SceneSignals.h
#	Tools/ResourceEditor/Classes/Qt/Scene/SceneTabWidget.cpp
#	Tools/ResourceEditor/Classes/Qt/Scene/SceneTabWidget.h
#	Tools/ResourceEditor/Classes/Qt/main.cpp


Former-commit-id: 7d5d9eba8d049f1f287c5c3bbd090eb7c327fe24

+++ Sources/Internal/Render/RHI/Common/MCPP/mcpp_main.cpp
@@ -38,10 +38,10 @@
 #include "_mcpp.h"
 
 #if PREPROCESSED /* Use "pre-preprocessed" header    */
-#include "mcpp.H"
+#include "mcpp.h"
 #else
-#include "system.H"
-#include "internal.H"
+#include "system.h"
+#include "internal.h"
 #endif
 
 /* Function pointer to expand_macro() functions.    */
@@ -773,7 +773,7 @@ static void mcpp_main(void)
                 }
                 if (keep_spaces && wrong_line && infile && *(infile->bptr) != '\n' && *(infile->bptr) != EOS)
                 {
-                    src_col = infile->bptr - infile->buffer;
+                    src_col = static_cast<int>(infile->bptr - infile->buffer);
                     /* Remember the current colums  */
                     break; /* Do sharp() now       */
                 }
